fix: Change PK datatype Integer to Long (#4)

+++ src/main/java/com/m1s/m1sserver/Repository/CounselSurveyRepository.java
@@ -4,5 +4,5 @@ import com.m1s.m1sserver.Model.CounselSurvey;
 import com.m1s.m1sserver.Model.RegisterSurvey;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CounselSurveyRepository extends JpaRepository<CounselSurvey, Integer> {
+public interface CounselSurveyRepository extends JpaRepository<CounselSurvey, Long> {
 }
