+++ src/main/java/com/m1s/m1sserver/Repository/InterestRepository.java
@@ -4,5 +4,5 @@ import com.m1s.m1sserver.Model.Interest;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.repository.CrudRepository;
 
-public interface InterestRepository extends JpaRepository<Interest, Integer> {
+public interface InterestRepository extends JpaRepository<Interest, Long> {
 }
