+++ src/main/java/com/m1s/m1sserver/Repository/RankingRepository.java
@@ -3,5 +3,5 @@ package com.m1s.m1sserver.Repository;
 import com.m1s.m1sserver.Model.Ranking;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface RankingRepository extends JpaRepository<Ranking, Integer> {
+public interface RankingRepository extends JpaRepository<Ranking, Long> {
 }
